implementation of uow and repository + two-tier arch
