+++ DataAccessLayer/Interfaces/IHaveId.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace WebUniversity.Models.Interfaces
+namespace DataAccessLayer.Interfaces
 {
     public interface IHaveId
     {
