+++ DataAccessLayer/Interfaces/IRepository.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace WebUniversity.Repository
+namespace DataAccessLayer.Interfaces
 {
     public interface IRepository<T> where T: class
     {
@@ -12,6 +12,5 @@ namespace WebUniversity.Repository
         void Create(T item);
         void Update(T item);
         void Delete(T item);
-        void Delete(int? id);
     }
 }
