+++ DataAccessLayer/Models/Course.cs
@@ -1,11 +1,8 @@
-﻿using System;
+﻿using DataAccessLayer.Interfaces;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
-using WebUniversity.Models.Interfaces;
 
-namespace WebUniversity.Models
+namespace DataAccessLayer.Models
 {
     public class Course : IHaveId
     {
