+++ DataAccessLayer/Models/Student.cs
@@ -1,11 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using DataAccessLayer.Interfaces;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
-using WebUniversity.Models.Interfaces;
 
-namespace WebUniversity.Models
+namespace DataAccessLayer.Models
 {
     public class Student : IHaveId
     {
