+++ WebUniversity/Controllers/CoursesController.cs
@@ -1,13 +1,10 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using DataAccessLayer.Interfaces;
+using DataAccessLayer.Models;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using SmartBreadcrumbs.Attributes;
-using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
-using WebUniversity.DataAccess;
-using WebUniversity.Models;
-using WebUniversity.Models.ViewModels;
+using WebUniversity.ViewModels;
 
 namespace WebUniversity.Controllers
 {
@@ -20,7 +17,7 @@ namespace WebUniversity.Controllers
         [Breadcrumb("Courses")]
         public override IActionResult Index(int page = 1, string search = null)
         {
-            IndexViewModel<Course> viewModel = Paginate(page, search);
+            IndexViewModel<Course> viewModel = HandleIndex(page, search);
 
             return View(viewModel);
         }
