--- WebUniversity/DataAccess/IUnitOfWork.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using WebUniversity.Repository;
-
-namespace WebUniversity.DataAccess
-{
-    public interface IUnitOfWork : IDisposable
-    {
-        IRepository<TEntity> GetRepository<TEntity>() where TEntity : class;
-        void SaveChanges();
-    }
-}
