+++ WebUniversity/TagHelpers/PageLinkTagHelper.cs
@@ -3,12 +3,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.AspNetCore.Mvc.Routing;
 using Microsoft.AspNetCore.Mvc.ViewFeatures;
 using Microsoft.AspNetCore.Razor.TagHelpers;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using WebUniversity.Models.ViewModels;
-using SmartBreadcrumbs;
+using WebUniversity.ViewModels;
 
 namespace WebUniversity.TagHelpers
 {
