+++ WebUniversity/ViewModels/IndexViewModel.cs
@@ -1,9 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
-namespace WebUniversity.Models.ViewModels
+namespace WebUniversity.ViewModels
 {
     public class IndexViewModel<T>
     {
