+++ WebUniversity/ViewModels/PageViewModel.cs
@@ -1,9 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace WebUniversity.Models.ViewModels
+namespace WebUniversity.ViewModels
 {
     public class PageViewModel
     {
