+++ WebUniversity/Views/Groups/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@using WebUniversity.Models.ViewModels;
+﻿@using WebUniversity.ViewModels;
 @model IndexViewModel<Group>;
 
 @{
