+++ WebUniversity/Views/_ViewImports.cshtml
@@ -1,4 +1,4 @@
-﻿@using WebUniversity.Models;
+﻿@using DataAccessLayer.Models;
 @using Microsoft.AspNetCore.Mvc.Localization;
 @addTagHelper *, WebUniversity
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
